[skip ci] Encapsulate lists of agents into HeartBeatInspector.kt
